// frontend/gear-vault-web/src/api/generated/client/models/RegisterWalkInCustomerDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CreateWalkInVehicleDto } from './CreateWalkInVehicleDto';
export type RegisterWalkInCustomerDto = {
    name?: string | null;
    emailAddress?: string | null;
    phoneNumber?: string | null;
    address?: string | null;
    vehicles?: Array<CreateWalkInVehicleDto> | null;
};