// frontend/gear-vault-web/src/api/generated/client/models/CreateWalkInVehicleDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FuelType } from './FuelType';
export type CreateWalkInVehicleDto = {
    vehicleNumber?: string | null;
    licenseNumber?: string | null;
    make?: string | null;
    model?: string | null;
    year?: number;
    fuelType?: FuelType;
};